requirements.txt
pandas
anthropic
python-dotenv
boto3
sendgrid
streamlit